+++ src/Components/ContentElement/PanelSingleElement.php
@@ -6,7 +6,7 @@
  * @package    contao-bootstrap
  * @subpackage Panel
  * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2018 netzmacht David Molineus. All rights reserved.
+ * @copyright  2014-2019 netzmacht David Molineus. All rights reserved.
  * @license    https://github.com/contao-bootstrap/panel/blob/master/LICENSE LGPL 3.0-or-later
  * @filesource
  */
@@ -14,6 +14,20 @@
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Panel\Components\ContentElement;
+
+use Contao\Controller;
+use Contao\Database\Result;
+use Contao\FilesModel;
+use Contao\FrontendTemplate;
+use Contao\Model;
+use Contao\Model\Collection;
+use ContaoBootstrap\Core\Helper\ColorRotate;
+use Netzmacht\Contao\Toolkit\Data\Model\RepositoryManager;
+use Netzmacht\Contao\Toolkit\Exception\InvalidArgumentException;
+use Netzmacht\Contao\Toolkit\Routing\RequestScopeMatcher;
+use Symfony\Component\Templating\EngineInterface as TemplateEngine;
+use function array_merge;
+use function is_file;
 
 /**
  * Class PanelGroupStartElement
@@ -26,4 +40,61 @@ final class PanelSingleElement extends AbstractPanelStartElement
      * @var string
      */
     protected $templateName = 'ce_bs_panel_single';
+
+    /**
+     * Repository manager.
+     *
+     * @var RepositoryManager
+     */
+    private $repositoryManager;
+
+    /**
+     * AbstractContentElement constructor.
+     *
+     * @param Model|Collection|Result $model             Object model or result.
+     * @param TemplateEngine          $templateEngine    Template engine.
+     * @param ColorRotate             $colorRotate       ColorRotate helper.
+     * @param RequestScopeMatcher     $scopeMatcher      Request scope matcher.
+     * @param RepositoryManager       $repositoryManager Repository manager.
+     * @param string                  $column            Column.
+     *
+     * @throws InvalidArgumentException When model does not have a row method.
+     */
+    public function __construct(
+        $model,
+        TemplateEngine $templateEngine,
+        ColorRotate $colorRotate,
+        RequestScopeMatcher $scopeMatcher,
+        RepositoryManager $repositoryManager,
+        string $column = 'main'
+    ) {
+        parent::__construct($model, $templateEngine, $colorRotate, $scopeMatcher, $column);
+
+        $this->repositoryManager = $repositoryManager;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function prepareTemplateData(array $data) : array
+    {
+        $data = parent::prepareTemplateData($data);
+
+        // Add an image
+        if ($this->get('addImage') && $this->get('singleSRC') != '') {
+            $repository = $this->repositoryManager->getRepository(FilesModel::class);
+            $fileModel  = $repository->findByUuid($this->get('singleSRC'));
+
+            if ($fileModel !== null && is_file(TL_ROOT . '/' . $fileModel->path)) {
+                $data['singleSRC'] = $fileModel->path;
+
+                $template = new FrontendTemplate($this->templateName);
+                Controller::addImageToTemplate($template, $data, null, null, $fileModel);
+
+                $data = array_merge($data, $template->getData());
+            }
+        }
+
+        return $data;
+    }
 }